FEATURE - i18n: install / setup next-intl - ui: install daisyui
