+++ app/[locale]/blog/page.tsx
