+++ app/[locale]/developer/page.tsx
