+++ app/[locale]/teacher/page.tsx
