+++ middleware.ts
@@ -1,0 +1,16 @@
+import createMiddleware from 'next-intl/middleware';
+import { locales } from './navigation';
+
+export default createMiddleware({
+  // A list of all locales that are supported
+  locales,
+
+  // Used when no locale matches
+  defaultLocale: 'en',
+  localePrefix: 'always',
+});
+
+export const config = {
+  // Match only internationalized pathnames
+  matcher: ['/', '/(fr|en)/:path*'],
+};
