+++ next.config.js
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
+
+const withNextIntl = require('next-intl/plugin')('./i18n.ts');
+
 const nextConfig = {
   output: 'standalone',
-}
+};
 
-module.exports = nextConfig
+module.exports = withNextIntl(nextConfig);
